+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
-- First version of JWT Extension.
+- Second version of JWT Support.
+  - Contains JWT Fuzzer for fuzzing the JWT's present in the request.
+- First version of JWT Support.
   - Contains scanning rules for basic JWT related vulnerabilities.